GMXPY-68 Fix algorithm.
Note the remaining to-dos and resolve the first couple.

The original implementation started applying the potential as soon as
there are any windows of data. Since our histogram is initialized to
zeros, we can start applying the potential immediately.

We have also begun to fix the places where values were not properly
normalized (taking the mean instead of the sum).