GMXPY-64 reference coordinate
Allow an arbitrary number of reference coordinates (zero or more) to
specify a path between the two sites of interest. Minimum image
convention is applied between each successive pair of sites in the
chain, but the overall vector between the first and last site can exceed
half a box length without being wrapped.