Resolve an unsafe lambda capture.
At least in Python 3, the py::str(std::string s) behavior does not seem
to result in a reference-counted copy of the provided string and results
in a double-free problem on at least some platforms when the then-owner
of the function object is destroyed (it outlives the original string).

To resolve, we copy the data member string to a local (non-member)
const std::string, which the lambda function seems to capture
appropriately and is properly cleaned up. The lambda function is defined
in a translation unit that has access to py::str(), so we're fine there.