Remove pre-rendered content from `docs/_build`.
To avoid bloating the repository and littering the `master` branch,
we will publish the rendered html from the `gh-pages` branch. This
commit removes the contents of `docs/_build` from the `master` branch.

The `docs/index.html` file can be removed in a follow-up commit
once we have automated curation of the `gh-pages` branch.

See issue #35.